+++ Sources/ParseSwift/Authentication/3rd Party/ParseLDAP.swift
@@ -80,7 +80,7 @@ public extension ParseLDAP {
         guard AuthenticationKeys.id.verifyMandatoryKeys(authData: authData) else {
             callbackQueue.async {
                 completion(.failure(.init(code: .unknownError,
-                                          message: "Should have authData in consisting of keys \"id\" and \"token\".")))
+                                          message: "Should have authData in consisting of keys \"id\" and \"password\".")))
             }
             return
         }
@@ -153,7 +153,7 @@ public extension ParseLDAP {
               completion: @escaping (Result<AuthenticatedUser, ParseError>) -> Void) {
         guard AuthenticationKeys.id.verifyMandatoryKeys(authData: authData) else {
             let error = ParseError(code: .unknownError,
-                                   message: "Should have authData in consisting of keys \"id\" and \"token\".")
+                                   message: "Should have authData in consisting of keys \"id\" and \"password\".")
             callbackQueue.async {
                 completion(.failure(error))
             }